Add files via upload
What is the optimal move for A?

Very simple approach:
A and B are grid searchers.
D is a greedy (going for the max possible area)
C is where most of the logic happens (i.e. C judges the state of the board and how D will react before choosing).

D often ends up with multiple possible choices, A and B take this into account by checking the minimum area they can guarentee regardless of D's choice.